refactor: asset diff capabilities
